replace old modal component (on presentation too), add size prop, fix some mobile styles
