+++ src/components/ui/modals/Modals.vue
@@ -4,16 +4,16 @@
       <div class="flex md12">
         <vuestic-widget class="modals-list larger-padding"
                         :header-text="$t('modal.title')">
-          <button class="btn btn-danger" @click="showSmallModal()">
+          <button class="btn btn-danger" @click="isShowSmallModal = true">
             {{'modal.small' | translate }}
           </button>
-          <button class="btn btn-info" @click="showMediumModal()">
+          <button class="btn btn-info" @click="isShowMediumModal = true">
             {{'modal.medium' | translate }}
           </button>
-          <button class="btn btn-warning" @click="showLargeModal()">
+          <button class="btn btn-warning" @click="isShowLargeModal = true">
             {{'modal.large' | translate }}
           </button>
-          <button class="btn btn-success" @click="showStaticModal()">
+          <button class="btn btn-success" @click="isShowStaticModal = true">
             {{'modal.static' | translate }}
           </button>
         </vuestic-widget>
@@ -21,61 +21,59 @@
     </div>
 
     <!--//Modals-->
-    <vuestic-modal :show.sync="show" ref="smallModal" v-bind:small="true"
-                   :cancelClass="'none'"
-                   :okText="'modal.confirm' | translate"
-                   :cancelText="'modal.cancel' | translate">
-      <div slot="title">{{'modal.smallTitle' | translate}}</div>
-      <div>
-        There are three species of zebras: the plains zebra, the mountain zebra
+    <vuestic-modal
+      v-model="isShowSmallModal"
+      size="small"
+      :title="'modal.smallTitle' | translate"
+      message="There are three species of zebras: the plains zebra, the mountain zebra
+      and the Grévy's zebra. The plains zebra
+      and the mountain zebra belong to the subgenus Hippotigris, but Grévy's
+      zebra is the sole species of subgenus
+      Dolichohippus. The latter resembles an ass, to which it is closely
+      related, while the former two are more
+      horse-like. All three belong to the genus Equus, along with other living
+      equids."
+      :okText="'modal.confirm' | translate"
+      :cancelText="'modal.cancel' | translate"
+    />
+    <vuestic-modal
+      v-model="isShowMediumModal"
+      :title="'modal.mediumTitle' | translate"
+      :okText="'modal.confirm' | translate"
+      :cancelText="'modal.cancel' | translate"
+      message="There are three species of zebras: the plains zebra, the mountain zebra
         and the Grévy's zebra. The plains zebra
         and the mountain zebra belong to the subgenus Hippotigris, but Grévy's
         zebra is the sole species of subgenus
         Dolichohippus. The latter resembles an ass, to which it is closely
         related, while the former two are more
         horse-like. All three belong to the genus Equus, along with other living
-        equids.
-      </div>
-    </vuestic-modal>
-    <vuestic-modal :show.sync="show" ref="mediumModal"
-                   :okText="'modal.confirm' | translate"
-                   :cancelText="'modal.cancel' | translate">
-      <div slot="title">{{'modal.mediumTitle' | translate}}</div>
-      <div>
-        There are three species of zebras: the plains zebra, the mountain zebra
+        equids."
+    />
+    <vuestic-modal
+      v-model="isShowLargeModal"
+      size="large"
+      :title="'modal.largeTitle' | translate"
+      message="There are three species of zebras: the plains zebra, the mountain zebra
         and the Grévy's zebra. The plains zebra
         and the mountain zebra belong to the subgenus Hippotigris, but Grévy's
         zebra is the sole species of subgenus
         Dolichohippus. The latter resembles an ass, to which it is closely
         related, while the former two are more
         horse-like. All three belong to the genus Equus, along with other living
-        equids.
-      </div>
-    </vuestic-modal>
-    <vuestic-modal :show.sync="show" v-bind:large="true" ref="largeModal"
-                   :okText="'modal.confirm' | translate"
-                   :cancelText="'modal.cancel' | translate">
-      <div slot="title">{{'modal.largeTitle' | translate}}</div>
-      <div>
-        There are three species of zebras: the plains zebra, the mountain zebra
-        and the Grévy's zebra. The plains zebra
-        and the mountain zebra belong to the subgenus Hippotigris, but Grévy's
-        zebra is the sole species of subgenus
-        Dolichohippus. The latter resembles an ass, to which it is closely
-        related, while the former two are more
-        horse-like. All three belong to the genus Equus, along with other living
-        equids.
-      </div>
-    </vuestic-modal>
-    <vuestic-modal :show.sync="show"
-                   v-bind:small="true" v-bind:force="true" ref="staticModal"
-                   :cancelClass="'none'"
-                   :okText="'modal.close' | translate">
-      <div slot="title">{{'modal.staticTitle' | translate}}</div>
-      <div>
-        {{'modal.staticMessage' | translate}}
-      </div>
-    </vuestic-modal>
+        equids."
+      :okText="'modal.confirm' | translate"
+      :cancelText="'modal.cancel' | translate"
+    />
+    <vuestic-modal
+      v-model="isShowStaticModal"
+      size="small"
+      :title="'modal.staticTitle' | translate"
+      :cancelClass="'none'"
+      :okText="'modal.close' | translate"
+      :message="'modal.staticMessage' | translate"
+      :noOutsideDismiss="true"
+    />
   </div>
 </template>
 
@@ -86,23 +84,12 @@ export default {
   data () {
     return {
       show: true,
+      isShowSmallModal: false,
+      isShowMediumModal: false,
+      isShowLargeModal: false,
+      isShowStaticModal: false
     }
-  },
-
-  methods: {
-    showSmallModal () {
-      this.$refs.smallModal.open()
-    },
-    showMediumModal () {
-      this.$refs.mediumModal.open()
-    },
-    showLargeModal () {
-      this.$refs.largeModal.open()
-    },
-    showStaticModal () {
-      this.$refs.staticModal.open()
-    },
-  },
+  }
 }
 </script>
