+++ src/vuestic-theme/vuestic-components/vuestic-modal/VuesticModal.demo.vue
@@ -11,19 +11,26 @@
       <vuestic-radio-button value="bottom" v-model="modalPosition" option="bottom"/>
       <vuestic-radio-button value="left" v-model="modalPosition" option="left"/>
     </div>
-    <vuestic-modal-new
+    <p>modal size</p>
+    <div>
+      <vuestic-radio-button value="small" v-model="modalSize" option="small"/>
+      <vuestic-radio-button value="medium" v-model="modalSize" option="medium" label="medium (default)"/>
+      <vuestic-radio-button value="large" v-model="modalSize" option="large"/>
+    </div>
+    <vuestic-modal
       v-model="isModalShow"
       title="Simple Popup, Full Width"
       message="About 64% of all on-line teens say that do things online that they wouldn’t want their parents to know about.   11% of all adult internet users visit dating websites and spend their time Some of the classify their behavior as “cyber affair”   More then 60% of employees use company PC for the personal needs during their work hours as long as 80 minutes per day."
       okText="Accept"
       cancelText="Decline"
       :position="modalPosition"
+      :size="modalSize"
     />
   </div>
   <div class="demo-container__item">
     <p>fullscreen modal with, slots use</p>
     <button @click="triggerFullscreenModal">Show modal</button>
-    <vuestic-modal-new
+    <vuestic-modal
       v-model="isFullScreenModalShow"
       :fullscreen="true">
       <template slot="header">
@@ -32,12 +39,12 @@
       <slot>
         <div>About 64% of all on-line teens say that do things online that they wouldn’t want their parents to know about.   11% of all adult internet users visit dating websites and spend their time Some of the classify their behavior as “cyber affair”   More then 60% of employees use company PC for the personal needs during their work hours as long as 80 minutes per day.</div>
       </slot>
-    </vuestic-modal-new>
+    </vuestic-modal>
   </div>
   <div class="demo-container__item">
     <p>no-outside-dismiss, no-esc-dismiss, max-width: 800px, max-height: 250px, with scroll</p>
     <button @click="triggerActionsModal">Show modal</button>
-    <vuestic-modal-new
+    <vuestic-modal
       v-model="isActionsModalShow"
       :no-outside-dismiss="true"
       :no-esc-dismiss="true"
@@ -50,7 +57,7 @@
   <div class="demo-container__item">
     <p>fixed layout && onOk, onCancel functions</p>
     <button @click="triggerFixedLayoutModal">Show modal</button>
-    <vuestic-modal-new
+    <vuestic-modal
       v-model="isFixedLayoutModalShow"
       max-width="800px"
       max-height="250px"
@@ -60,24 +67,25 @@
       message="About 64% of all on-line teens say that do things online that they wouldn’t want their parents to know about.   11% of all adult internet users visit dating websites and spend their time Some of the classify their behavior as “cyber affair”   More then 60% of employees use company PC for the personal needs during their work hours as long as 80 minutes per day. About 64% of all on-line teens say that do things online that they wouldn’t want their parents to know about.   11% of all adult internet users visit dating websites and spend their time Some of the classify their behavior as “cyber affair”   More then 60% of employees use company PC for the personal needs during their work hours as long as 80 minutes per day."
     >
       <template slot="header"><p>Step 2. Centered Layout</p></template>
-    </vuestic-modal-new>
+    </vuestic-modal>
   </div>
 </div>
 </template>
 
 <script>
-import VuesticModalNew from './VuesticModalNew'
+import VuesticModal from './VuesticModal'
 import VuesticRadioButton from '../vuestic-radio-button/VuesticRadioButton'
 
 export default {
-  components: { VuesticRadioButton, VuesticModalNew },
+  components: { VuesticRadioButton, VuesticModal },
   data () {
     return {
       isModalShow: false,
       isFullScreenModalShow: false,
       isActionsModalShow: false,
       isFixedLayoutModalShow: false,
-      modalPosition: 'center'
+      modalPosition: 'center',
+      modalSize: 'medium'
     }
   },
   methods: {
